update_reminders_requirements.txt
google-cloud-scheduler>=1.2.1,<2.0
PyYAML>=5.2,<6.0
python-dateutil>=2.8.1,<3.0
